// src/client/toteco/models/recover-account.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface RecoverAccount
 */
export interface RecoverAccount {

    /**
     * @type {string}
     * @memberof RecoverAccount
     */
    id?: string;

    /**
     * @type {string}
     * @memberof RecoverAccount
     */
    username?: string;

    /**
     * @type {number}
     * @memberof RecoverAccount
     */
    recoveryCode?: number;
}