// src/client/toteco/models/update-password.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface UpdatePassword
 */
export interface UpdatePassword {

    /**
     * @type {string}
     * @memberof UpdatePassword
     */
    id?: string;

    /**
     * @type {string}
     * @memberof UpdatePassword
     */
    password?: string;
}