// src/client/toteco/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface UserDTO
 */
export interface UserDTO {

    /**
     * @type {string}
     * @memberof UserDTO
     */
    username?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    name?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    surname?: string;

    /**
     * @type {number}
     * @memberof UserDTO
     */
    birthDate?: number;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    email?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    password?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    photo?: string;

    /**
     * @type {string}
     * @memberof UserDTO
     */
    role?: string;
}